refac(api): refactor & cleanup of API
