+++ cooking-user/sj2324-5ehif-cooking-user.Webapi/Services/PasswordUtils.cs
@@ -7,9 +7,9 @@ public class PasswordUtils : IPasswordUtils
 {
     public string HashPassword(string password)
     {
-        using (SHA256 sha256 = SHA256.Create())
+        using (var sha256 = SHA256.Create())
         {
-            byte[] hashedBytes = sha256.ComputeHash(Encoding.UTF8.GetBytes(password));
+            var hashedBytes = sha256.ComputeHash(Encoding.UTF8.GetBytes(password));
             return BitConverter.ToString(hashedBytes).Replace("-", "").ToLower();
         }
     }
